revert previous change
On further investigation, it appears that right clicking anywhere not on the item label or icon causes the global menu to appear (with the previous change).  This is due to wxPython's HitTest method always return (wx.NOT_FOUND,wx.LIST_HITTEST_NOWHERE) when right clicking somewhere not on the item's text or icon.  Unfortunately, this means we cannot detect the difference between right clicking on say, the "Order" column's text, and the scrollbar.

git-svn-id: svn://svn.code.sf.net/p/oblivionworks/code/Programs/Wrye%20Bash@2256 c80116c5-1073-4541-b2a9-63b9fd661f84