BOSS API usage fixes: if no masterlist.txt is present, load taglist.txt instead.  If neither is present, present the user with an error (taglist.txt come with Wrye Bash).
One bug I found - BOSS API doesn't internally create a Bash Tag map if the passed masterlist.txt is empty - so further calls to GetBashTagMap will return an error code instead of an empty tag map. - report to wrinklyninja, so in the event that taglist.txt is missing, I can create a blank one for use with the BOSS API and not encounter this error.

git-svn-id: svn://svn.code.sf.net/p/oblivionworks/code/Programs/Wrye%20Bash@2320 c80116c5-1073-4541-b2a9-63b9fd661f84