Split game.name into .displayName and .fsName.
For #144. Also moved the default ini string into the game definition
files.

As it is now, some UI elements may use the fsName string where the UI
and filesystem interactions are tightly-bound (eg. bush.setGame,
bush.detectGames), but that will only look odd for the Fallout games,
and can be corrected as and when it needs to be.