FIX for Screens/People main menu help:
Under #174.
People and Screens main menus did not display help items. Reason: their last
link was a submenu and its items set the Link.Popup attribute to the submenu.
Now Link.Popup is set once in Links.PopupMenu - should be final.

	# This is the 2nd commit message:

Link.Popup held a ref to menu - to the best of my knowledge this was a leak.