RIP: PatcherDialog - changes in otherPatcherDict:
Important: otherPatcherDict changed from a 'string' -> Patcher dictionary
to a 'string' -> 'string' one.

-otherPatcherDict = {
-    AliasesPatcher().__class__.__name__ : CBash_AliasesPatcher(),
+otherPatcherDict = {
+    'AliasesPatcher' : 'CBash_AliasesPatcher',

Was only used in PatchDialog.ConvertConfig
and there only the name of the stored patcher class was used:

-    def ConvertConfig(self,patchConfigs):
-        newConfig = {}
-        for key in patchConfigs:
-            if key in otherPatcherDict:
-                newConfig[otherPatcherDict[key].__class__.__name__] = patchConfigs[key]
-            else:
-                newConfig[key] = patchConfigs[key]
-        return newConfig
-

Turned to:

+    def ConvertConfig(self,patchConfigs):
+        newConfig = {}
+        for key in patchConfigs:
+            if key in otherPatcherDict:
+                newConfig[otherPatcherDict[key]] = patchConfigs[key]
+            else:
+                newConfig[key] = patchConfigs[key]
+        return newConfig

The old form would lead to weird cyclic imports plus it did not buy us
anything at all - it was the same as hardcoding the strings. Still a
hack but at least it does not look like a centipede anymore.

Wrong except order corrected - when cancelling was displaying an error
dialog.

In mod_links the isinstance checks are now again with the 'bosh' version
of the patchers (see for instance basher line 14292 at
ef4c901fa1c2495704b9abd6f303422f15cac66a).

Changed the style of the imports to explicitly specify the patcher module:

-class UpdateReferences(UpdateReferences,ListPatcher): pass
+class UpdateReferences(base.UpdateReferences,ListPatcher): pass

The from ... import XXX form was too confusing (and naturally produced
warnings)