configIsCBash moved to patcher.__init__:
I wanted to rip as much out of bosh as I could. Turns out ModInfos uses
the isMergeable stuff which in turns uses modInfos global. I could
go with in-function imports but preferred to not sweep this under the
carpet: bosh needs splitting, no escaping from this (and mergeability
stuff stays there to remind us).

configIsCBash() is another beast: the iron rule is "no module imports
basher" (apart bash). configIsCBash() belongs to the (GUI) configs
but those are used outside of basher. Stuck it right in the top of
patcher for now, for everyone to see.

Closes #3, at very long last.

Concerns #163, #178 also - we must decide on GUI patchers config handling.