ListCtrl.ReorderItems() - Tank works:
This opens the way to transfer Sorting to UIList - these dicts used in Tank
is apparently the only way to implement sorting in the control:
http://wiki.wxpython.org/Recipes%20L10n%20I18n%20-%20How%20to%20sort%20ListCtrl%20items%20under%20a%20locale.
Mopy/bash/basher/__init__.py:
addMarker would error as GetIndex() would not find the u'====' key
in _itemId_item dict - will be inserted in RefreshUIMods.
Took the opportunity to prune a self.items usage in OnBeginEditLabel
Mopy/bash/bosh.py:
transferred the refresh inside the data method and pruned a self.data
access - DataDict is a dict after all and the data attribute should be
eliminated eventually.