Nitpick in Installers sort - TankData param API deprecation:
Files, size = -1 so that markers sort always after files (with 0 length)
Installers markers modification time is now shown (it was taken into
account when sorting anyway).

The idea to have a 'data source' is not bad ofc - but the way it was
implemented ended in an unmaintainable mess. So first unify everything
under UIList and then we'll see.
Plus inheritance and generality are NOT always the Right Thing. Those
are singleton classes. For an example where inheritance and generality
_is_ the right thing see the Link API.

Signed-off-by: MrD <the.ubik@gmail.com>