Revert previous commit - note on default settings and events:
Concerns #178.
This reverts commit 30c26b49bd8fc9235eb6ec7c0e30a2e6f81c0f45 and recreates
its effect by a hacky workaround. The change in the aforementioned commit
introduced a regression in installers tab, namely users who had installers
enabled saw them suddenly disabled (see #187).

This is all due to settings not being saved if they are not changed from
the _harcoded_ default. I deeply disliked this design decision and here is
the proof. To save a couple KB from the settings file one:

- is not able to change these harcoded defaults
- has to mark setChanged which needlessly (and greatly) complicates the code
- gets key errors if the default is missing or has to use 'setDefault' and co

Centralizing defaults is fine and will help with settings refactoring but
it should not affect in any way the rest of the code, let alone reach the user.

Incidentally this and previous commits holds the key to the
wx.EVT_MOUSE_CAPTURE_LOST - see:
http://sourceforge.net/p/oblivionworks/bugs/251/

The effect of having to click two times is because there are _two_ dialogs,
the first on clicking on the installers tab while firstRun is true,
the second on clicking OK whereup the RefreshData() callback is triggered,
which in turn call installers OnShowPage.
This may be also be related to the installers tab 'sticking' on the mouse
pointerafter the dialog.
This hack should address the wx.EVT_MOUSE_CAPTURE_LOST, will be removed
soon.

EDIT: first (succesful) attempt was:
@@ -2950,3 +2950,5 @@ def ShowPanel(self, canCancel=True):
         # TODO(ut): refactor, self.refreshing set to True once, extract methods
+        if settings.get('bash.installers.lock', False): return
         if settings.get('bash.installers.isFirstRun',True):
+            settings['bash.installers.lock'] = True
             # I have no idea why this is necessary but if the
@@ -2966,2 +2968,3 @@ def ShowPanel(self, canCancel=True):
             settings['bash.installers.enabled'] = balt.askYes(self,fill(message,80),self.data.title)
+            del settings['bash.installers.lock']
         if not settings['bash.installers.enabled']: return

The way I did it finally is much more optimal (RefreshData()is prevented
from running) and will also serve us well in #123.