Oblivion esm version swapping fixups:
Closes #198.

The 305 bug was due to _Default.Execute() lines:

-            BashFrame.modList.RefreshUI()
-            Link.Frame.RefreshData()

Same lines were in correct order in _ProfileLink - so essentially another
copy paste bug. Took the opportunity to add help to the profiles links.

setProfile=True named argument - not sure I like the different behavior
of saves and mods tabs - documented here:
http://wrye-bash.github.io/docs/Wrye%20Bash%20Advanced%20Readme.html#tools-swapping

The assertions were due to me not clearing the internal dicts when deleting
all items from the _gList - essentially the key error below was not thrown:

    class _ProfileLink(CheckLink):
        def Execute(self, event):
            #....
            Saves_Profiles.swapOblivionVersion(newSaves)
            Link.Frame.SetTitle()
            self.window.details.SetFile(None) # <------ DeleteAllItems()
            Link.Frame.RefreshData()

            # PopulateItem() ....
            fileName = item
            try:
                itemDex = self.GetIndex(item)
            except KeyError:
                itemDex = self._gList.GetItemCount() # insert at the end
                insert = True

so it was not inserting - just trying to update the "existing" list item
`self._gList.SetStringItem(itemDex, colDex, labels[col])`.
Alpha is there for a reason turns out :)

Renamed DeleteAllItems() to DeleteAll() to differentiate it from the
DeleteAllItems ListCtrl method.