DataDict.delete():
A hack suited to current delete()s.

minuses:

- itemOrItems - Yak !
- **kwargs: always a bad idea (exceptions apply). Notice for instance how
 was askOk=False FileInfos default being constantly overridden.
- Table, Settings, MelQustLoaders and MelCnamLoaders DataDict subclasses
did not have a delete implementation. I do not want delete start being used
in those classes - kind of enforced by raising NotImplementedError.

plus:
an API is formed in basher DeleteItems - the fact that this is fed to
a kwargs is a minus as noted. Closes #32, as using this API we can define
recycle policies, etc.

Signed-off-by: MrD <the.ubik@gmail.com>