Interlude: shellMove default arguments flip, fileOperation nitpick:
The parameters in fileOperation are evil with pure MS evil:
https://msdn.microsoft.com/en-us/library/windows/desktop/bb759795%28v=vs.85%29.aspx
See also Tim Golden stuff:
http://winshell.readthedocs.org/en/latest/file-operations.html
The isinstance calls I tried to simplify as seen.
Added silent=False parameter to shellMove which permitted me to get rid
of fileOperation in bosh:
-        balt.fileOperation(balt.FO_MOVE,tempFile,dest,False,False,False,True,None)
+        balt.shellMove(tempFile, dest, askOverwrite=True, silent=True)

This permitted me to make _fileOperation balt-private.
Why askOverwrite=True (as it was noConfirm=not askOveright =False
but askOveright is now False by default so I make it True as it used to be)
escapes me but does not harm.
_fileOperation should be hidden away in a library (and finished, probably
breaks WINE ?). Exposing a programmer to _fileOperation may result from
couple days loss of time to permanent inability to program.
A dictionary example of the usefulness of keyword args.
TODO: noConfirm=not askOk_ - as in delete(...recycle=not dontRecycle...).
Took the opportunity to set the default in Path.tempDir(prefix) - all calls
used prefix=u'WryeBash_'.

 # This is the 2nd commit message:

SSS _fileOperation:

Note I changed the constants to match the shellcon definitions and that I
moved the `parent = parent.GetHandle() if parent else None` inside the
`if shell is not None` branch - it would make `if not askYes(parent,...`
blow in the else branch.

For the code for shell.pyd see:
http://stackoverflow.com/q/29053189/281545