Mismatch between plugins.txt and loadorder.txt (Skyrim) - impasse:
Re:

-            if self.selected != modInfos.getOrdered(self.selected,False):
-                modInfos.plugins.saveLoadOrder()
-                self.selected = modInfos.getOrdered(self.selected,False)
-                deprint("Mismatched Load Order Corrected")

The modInfos.plugins.saveLoadOrder() would in the case of Skyrim end
up saving the plugins.txt also along loadorder.txt. So what the deleted
code above is doing is: if the order of existing `plugins.selected` is not
the same as in newly loaded load order resave load order (so also
plugins.txt) and reorder `plugins.selected`. So far so good.
Till commit 72304257fa77ef6f43c8d8787651d8b70a060250 which includes this
change in Plugins.refresh():

@@ -3084,4 +3084,4 @@ def refresh(self,forceRefresh=False):
         if hasChanged or forceRefresh:
-            self.loadActive()
             self.loadLoadOrder()
+            self.loadActive(lo_with_corrected_master=self.LoadOrder)

So till then loadLoadOrder would be getting a freshly loaded `.selected`
while with the commit above the code subtly broke.
However the code was already subtly broken: loadActive() would order
the `selected` with a newly loaded load order as seen in
liblo.GetActivePlugins which calls liblo.GetOrdered() so the line

if self.selected != modInfos.getOrdered(self.selected,False):

would never evaluate to True. Verify this in 305 by manually reordering
plugins.txt in Skyrim and noticing Bash will never edit it. So basically
Bash displayed the order in loadorder.txt while the game used the one in
plugins.txt. Boom.
I moved the code that fixes the mismatch in loadActive and removed reordering
of the plugins.txt returned by liblo (by binning liblo.GetOrdered()). Real
fix is redefining data structures for active and loadOrder (as seen in
load_order.py).

But wait - liblo keeps active internally as a set ! Does not return the
plugins.txt order ! So all these mechanisms are for naught - the code was
calling:

    -                modInfos.plugins.saveLoadOrder()

cause this method indeed syncs plugins.txt with loadorder.txt but it was
unable to detect a mismatch

C++ time, or just read the plugins.txt in python ?