liblo.py Nitpick:
Just the beginning of setting up an internal, solid load order API. Most
of the work here is transitional - refactoring the existing API in place
to safely replace it.

- globals() is evil - evil, I say. Introduced static scoping to be later
improved.
- Removed non WB specific code - confuses import detection which has been
a constant annoyance while I was refactoring - similar code in cint is far
more annoying, but this beast I won't yet touch. Plus I want
centralized control for unicode stuff - so I removed _uni() and let GPath
handle that. (Note that GPath and _uni calls were 1 to 1 and always chained -
maybe not a performance issue, certainly a readability one, especially for
people new to the code).
- Boiler plate code in list overrides removed - one of them should probably
be a set override. A refactoring exercise - ideally will be confined
in my branch. For the curious:

 Access base class attribute in derived class - in “class scope”:
http://stackoverflow.com/q/28734919/281545

 Override list's builtins dynamically in class scope:
http://stackoverflow.com/q/29858405/281545

 I do not agree with overriding list - too complex - see
 72304257fa77ef6f43c8d8787651d8b70a060250 and ex. (Turns out most of it was
 unused anyway (another indication of complexity).)
- Added `assert super_count in {0, 1}` in count() to get feedback for #100.
 (Turns out it's dead code anyway.)
- Fixed 'active' unused var, LibbsaVersionError.

Much more work is needed to specify an API, underscore the rest, move
it to a package and write a python layer cache.