Bye ActivePluginsList - never used outside liblo:
The only method that returned an ActivePluginsList instance was
_GetActivePluggins, the getter of ActivePlugins property. This property
was only used in ActivePluginsList.ReSync(). Bottom line:
ActivePluginsList was unused. Appart from being complex with its _DB and
ReSync() and all, it was setting _and_ getting stuff from liblo - that is
two system calls instead of one. The switches from c to python mode
should be kept to a minimum - doing what is to be done in python and
delegate to C once.

Couple newlines removed.