ModInfos.isActiveCached():
Most of the time Bash needs to know if a mod is active according to the
"last good load order" - and that is self.plugins.lord.
Revisiting all isSelected uses - note this:

@@ -4591,5 +4591,5 @@ def unselect(self,fileName,doSave=True):
         #--Unselect children
-        for selFile in self.plugins.selected[:]:
+        for selFile in self.plugins.selected[:]: # we recursively manipulate it
             #--Already unselected or missing?
-            if not self.isSelected(selFile) or selFile not in self.data:
+            if not selFile in self.plugins.selected or selFile not in self.data:
                 continue

In this case the rather awkward named isActiveCached justifies its name.
Now to ModInfos.ordered.
Bye isSelected() - of note that isActiveCached is O(1).