Performance fixup for tabbing in and out of Bash - needs work:
Under #123

The fix is ofc the marked one (!) but this lead to a new parameter to
refresh() - as now we do not rely on liblo to see if anything changed
in the LO in Oblivion and co (we have the info on modification times)
The doInfos=False path however would not trigger a refresh on modification
time change as before (and plugins refresh would now return False) hence
_modTimesChange. Leading underscore stands for hack here, this particular
hack is used to speed up refresh when we know that no scanning of Data/
is needed nevertheless LO changed due to a modification time change
we performed.
Real fix is that Bash knows the lo when not using txt file - so must
systematically use this info instead of having liblo rescan all
modification times - another goal for refresh() refactoring (edit:
see 'Short circuit GetLoadOrder for Oblivion' commit)

Note _checkUncheckMod() calls just refreshInfoLists() instead of refresh()

The order of FileInfos.refresh(self) and self.canSetTimes() calls should
not matter.