ListBoxes eternal bug fixed - needs work:
'message' parameter was unused from day one

Based on:
http://wxpython-users.1045709.n5.nabble.com/Wrapping-StaticText-again-td2362691.html

I had to also change delete to display the (finally displayed) correct
message depending on if the operation is reversible or not.

Removed unused 'labels' code

TODO: The Freeze/Thaw combo does not seem to do much, namely the text
is visibly resized. Moreover AddGrowableRow results in odd line splitting
while the text is resized, the tooltip does not show on delinquent masters
etc (edit: last one fixed in next commit)

In general this hack and slash approach to wx should give way to a
documented and systematic way of building UIs.