compressionSettings() and bosh pack:
A lot of work on compressCommand() -> it would blow with my command line
(7zExtraCompressionArguments=-ms=4g -mx=9) when I selected zip compression.
Still a WIP - for instance compressCommand() must be extended to produce
the command in packToArchive() and all this stuff (along with writeExts
etc) be moved to an archives.py module.