DataDict.delete_Refresh() - I need to refresh from someplace central:
Under #172.
Introduced to prune a refreshLoadOrder(forceRefresh=True) in ModList -
that is _the UI_ - and to introduce a lighter ModInfos.refresh() for
deletions.
Ugly name for an ugly hack - note Tank/List differences still.
This will be addressed fully when Tank is no more and when refresh is
aware of deleted, added, modified - under #205.

Renamed _postDeleteRefresh to _postDeleteRefreshUI for (needed) clarity.

 # This is the 2nd commit message:

 Bye _postDeleteRefreshUI:

 Tracking of BAIN files belongs to the data model