_DetailsMixin and _EditableMixin:
Under #163.
_DetailsPanelMixin -> _DetailsViewMixin for clarity
API is still beta but getting there - basically SahPanel should stop
inheriting from  _DetailsViewMixin and Tank should bite the dust
and we're there.
Ah yes - and find a way to use super in NotebookPanel hierarchy (hard !)
But this is 307+

This commit fixes tracebacks due to differences in SetFile, DoCancel
etc (for instance note that the `if not self.modInfo: return` is missed
in SaveDetails.SetFile). Now that the code duplication dust settles is seen
that _SashDetailsPanel is a details panel specialized for mods and saves
(they both have master lists). More work is needed to make the API more
flexible - it will be tested when Installers, People etc have details
classes (instead of using theyselves).

self.top, self.bottom is just for clarity - might as well use left, right

Opens the way to fix the UI peculiarities in MasterList discussed in #192.