MasterList.allowEdit and cosaves warns:
I reworked the logic of the masterlist edits so the save and cancel
buttons are only enabled when there are edits (finalized in next commit
due to event mess and wx unix programmers). It's another thing to
allow edits and another to edit.
I decided to add a menu item to enable/disable editing.
Under #192 for cosaves warns and the reported erratic behavior when
clicking on the master list.
I got rid of DoItemMenu override (the only override of the base method)
and reworked OnLeftDown override - what we really needed was vetoing
OnBeginEditLabel. Plus the way it is now OnLeftDown Bash checks for renames
and refreshes the master list if there are any pending.
InitEdit would only show renames if the renamed file was not present - this
made no sense (I hope) - removed this and added bolding of renamed files -
was very difficult to see what was going on.

Mopy/bash/basher/mod_links.py: added self.window.SetMasterlistEdited()
as InitEdit() was doing this before - untested (the class is unused)

Signed-off-by: MrD <the.ubik@gmail.com>