Fix for _winreg not being available on linux:
Even with the import fix Bash would erroneously crash with
"No supported games found" - so edting was needed in bush. I ended up
importing _winreg once - WIP:

In bass - bass is amongst others a showcase of problematic module
bindings ;)
Refactoring WIP
In bass cause it's used by bosh _and_ basher

Mopy/bash/bosh.py:
The registry is used in windows to get user paths - hacked up some code to
the same effect for linux and should add to the linux readmes (when) that
people better use cli arguments for those.

Mopy/bash/basher/links.py:
windows only code - yak - links.py is on its way to be split, but this
buttons stuff is in the way

Mopy/bash/windows.py: is not guarded with `if winreg` as it will be blocked
in next commits in linux (btw this one causes the problems in WINE canVista
path).