Bugfix in _corruptedWarns():
Triggered by new Fallout 4 code. See:

https://github.com/wrye-bash/wrye-bash/commit/622f8c3b9c56a4fcf4e9fd6fbe286613b5132ca9

for a discussion - namely on what those invalid versions are:

> Invalid versions is a file format thing. In the TES4 header, there's a
version field, which really refers to the version of the file format.
Bethesda sets this essentially. Anything not conforming to what was
initially released (vanilla) or what the GECK outputs is indicative of
a plugin that was not made with official tools, and hence the warning.

Utumno:
Took this opportunity to simplify the code in ListBoxes taking advantage
of Path `__unicode__` method.