Fixups in __extractOmods - test:
- omodMoves, omodRemoves = set(), set() should be out of the loop
- progress(i, omod.stail)
- below:

                    -    omodMoves = [x for x in omodMoves]
                    +    omodMoves = [x for x in omodMoves if x.exists()]

- failedOmods omod.body -> omod.tail
Note that extractOmodsNeeded was checking the wrong thing (basically
omod.tail, as this is what was returned by dirs['installers'].list())
to unexplored consequences. It was more correct to patch adding in
failedOmods the tails than testing for the body.

Note also that `omod.body) for omod in omodMoves]` in:
-                        omodDests = [
-                            dirInstallersJoin(u'Bash', u'Failed OMODs',
-                                              omod.body) for omod in omodMoves]
-                        balt.shellMove(omodMoves, omodDests, self)
+                        _move_omods(omodMoves)

became :
+                dests = [dirInstallersJoin(u'Bash', u'Failed OMODs',
 omod.tail)  <----------------------- was body
+                         for omod in failed]

Extract methods it's called - aka DRY.

Test, if you still care for omods in 2015.