[!]Dropping smelly check from ModList.RefreshUI():
We should make sure that the files we require their presentation to be
refreshed are included in modInfos and are up to date (duh!).

See: 6452bcb4b44d4b434e8636b35b69351cb6742869

Exclamation mark cause you need to keep an eye open for tracebacks
due to this commit - though I doubt there will be any.