Updated MelCoed: RRR mention 282
The second Integer for MelCoed is a FormID if Owner is an NPC and thus
must be mapped.  If Owner is a Faction it is a signed integer not unsigned
belonging to a FormID as with an NPC.  Owner can be either type of FormID
but followed by two different types of integers.

See # 282