_RefreshingLink to encapsulate refreshDataSizeCrc - notes:
For a discussion of the points below see:
https://github.com/wrye-bash/wrye-bash/commit/8abc3cd8a9ef787b2310c8a637be3b383fb01eeb

- Installer_SkipRefresh I dropped installer.refreshDataSizeCrc call as
the last thing installer.refreshBasic(file) does is to call
refreshDataSizeCrc.

- Installer_SkipRefresh: Binned heuristics - takes no time even for
large projects - and toggling skipRefresh won't be performed that often,
plus there may be false negatives - quoting @lojack5:

- BAIN is all happy.
- Turn on Skip Refresh on an installer that is not installed. This is
important, because after executing an install on other packages, even if
Skip Refresh is selected, its status will still correctly be updated,
because BAIN knows it has files installed that match things that
changed.
- Install some stuff to the data folder, changing its contents. Here
it's important that the stuff that is changed is also files that the
installer would want to change.
- Turn off Skip Refresh on the installer. The installer hasn't changed,
so that check would make it so its status isn't updated based on the
new files in the Data folder.

TODO:

refreshBasic vs refreshDataSizeCrc vs refreshStatus ?

+    _refreshType = 'N' # so what's this about exactly ?

Why the difference in Installer_SkipVoices:

+    _refreshType = 'NS'

Note it did  not call installer.refreshStatus.

----

That's related to BAIN work (#219) but here cause collides with event
pruning stuff.