UIList.labels:
Under #163

This code is set it and forget it - now one can focus on the methods of
the classes as opposed to their internal mechanics.
Last piece of abstract code gone from UIList - as a bonus only the
labels for displayed columns are calculated (annoyed me since I
implemented getLabels()).
Profiling suggests that number of calls goes up if all columns are
enabled but time stays the same for practical purposes (fluctuates at
the order 0.01 sec).
I did not like my use of defaultdict - if the column did not exist it
should error - ofc this would be a dev error but still.
PopulateItem uses 'item' variable now - more accurate but needs
comments.
voCurrent in bosh.ModInfos where it belongs.
Deprecating '.allCols' read only setting - cleaner, note ugly
MasterList override gone.
Note I sacrificed a space to wrap at 79, wrapping is a metric a space
not.

107 insertions(+), 111 deletions(-)

Signed-off-by: MrD <the.ubik@gmail.com>