Installers rename refactoring WIP:
Under #219.

Given that self.data.pop(installer) is a noop as far as I can see and:

                    if InstallerType is bosh.InstallerMarker:
                        del data[archive]

only deletes old entry for markers - the archive is never removed from
data which results in refreshInstallers return True and all subsequent
methods run in irefresh(). I believe this was not meant to be so and
that actually irefresh is not needed - will remove it in
'renameInstaller returns a tuple to avoid refreshes' commit.