renameInstaller returns a tuple to avoid refreshes:
Under #219, #205.

See http://stackoverflow.com/a/33743452/281545 for tuple(any(zip))).
Note that mfiles, ifiles were generators so their bool values were True,
even if the generators were empty:
http://stackoverflow.com/q/661603/281545
Generators would buy us nothing here, the lists are a few items long
(ifiles usually empty). The ifiles one was especially slow (see #247).

I dropped the irefresh call - to the best of my understanding was
unneeded.
It would scan the installers dir and, as the old name was not removed
from self.data, it would detect something changed and trigger the rest
of the refreshes in irefresh - those were unneeded too IIUC.

If it was not unneeded that is bad design - heck I just renamed a
couple files - _nothing_ really changed.