Exception handling in installers/screens renames - WIP:
Under #259.

Fixes broken UI when an exception is thrown - the file would be left
displayed as whatever>?.rar/jpg (>,? are invalid chars) - ofc the
filename on disk was not changed.
TODO: factor out common code

Added a comment in balt and took wx stuff out of ScreenList.OnLabelEdited
Shadowing warns (now that the code is visited)