HitTest to balt.UIList._getItemClicked() - to skip or not to skip:
Under #190.
Note that in InstallersList.OnDClick event was not skipped `if not item`
Now it is skipped, for better or worse (seems to be for no difference).
On a closer look the event was sometimes skipped sometimes not - left
the mess intact (apart from InstallersList) but one of these days this
must be cleaned up.

ModList.OnLeftDown should be left alone though - the event should only
be skipped as it is - I wonder what's the purpose of the SetDnD call.