bosh/converters.py I - ripping converters stuff form InstallersData:
I introduced a new DataDict override: ConvertersData. It was a design
mistake to put installers and converters in the same class - dearly
paid as the BAIN classes are 1k line classes and naturally not
maintainable. Starting with this commit I will try to untangle them with
the final goal to rip a brand new Converters tab - see #222.

Passing directories as arguments of ConvertersData init() was done to
decouple bosh (bosh.dirs) from bosh.converters.py. Now that dirs will
be moved to bass this is not strictly necessary - I just keep it as an
alternate design we may consider - dirs is a coupling issue.

Minimal wrapping/formatting

Mopy/bash/basher/installer_links.py:
Renamed instData -> idata