Factoring out used once and only once logic in refreshSizeCrcDate():
I was staring at this code for long till I discerned that half of it
is used once (when scanning Data/ and we are on the top level).
Factored out to a method to fix the bug storm there - not an accident,
the code was unreadable.

Note newSDirs = (x for x ... SkipBAINRefresh) now an iterator - and
'\r\n' in log (see parent).