[!]FIX:Data directory refresh will now skip distantlod and more - notes:
Distant lod was meant to be skipped - so I went ahead and skipped some
of the other directories (cf _init_global_skips later on on this
branch). However as discussed in #219 this will result in BAIN
forgetting it has installed those files on a restart - so if a user
enables a skip and does not anneal all, on restarting BAIN won't add the
skipped files to data_sizeCrcDate and won't detect them as dirty (IIUC
of course since BAIN core is no less than complex).
Since distandlod (sic) was meant to be skipped I decided not to skip the
skips - which would be one way out of the problem, partial though as
docs are already skipped (and not later uninstalled, see report by
@alt3rn1ty in #219). Possible solutions:

- Force an anneal all when enabling a global skip
- refactor core BAIN uninstall to rescan skipped (horror)
- make full refresh ignore those global skips (not the silent ones
though) - this maybe recommended anyway - and have the user launch
a full refresh if he/she wants to completely uninstall

Signed-off-by: MrD <the.ubik@gmail.com>