REVIEW: skipEspmVoices to init_skips:
Maybe not as checked for everything (but only if there are
self.espmNots) ?
Workaround _ move `for lam in skips:` AFTER the

if fileExt in Installer._extensions_to_process: # process attributes
	if Installer._attributes_process[fileExt](self, fileLower,
                                                  full, fileExt):
		continue
if fileExt in global_skip_ext: continue # docs treated above
elif fileExt in Installer._executables_process: # and handle execs
	if Installer._executables_process[fileExt](
			checkOBSE, fileLower, full, archiveRoot, size, crc):
		continue

CHANGE: Note that skip espm voices will run even if overrideSkips==False
but not if skipVoices is set, as they would be skipped anyway.

Maybe we should add those to "noisy skips" ?

Under #270.