Moving silent skips to the beginning - REVIEW: see also parent:
Does:

	if fileEndsWith((u'thumbs.db',u'desktop.ini',u'config'))

need to be after `if type_ == 2: #--Complex archive` ?
(EDIT: no but the starts with test needs to be after the typpe==2 test
as fileLower may be remapped in there)

Track other manual usages of the same pattern - see refreshBasic:

if file.startswith(u'--'): continue

or:

            if item.s.lower().startswith((u'bash',u'--')): continue

in refreshInstallersNeeded() and

            if archive.s.lower().startswith((u'--',u'bash')): continue

in refreshInstallers().

I want this completely centralized.

Under #270.