iPanel.refreshed -> _data_dir_scanned and ShowPanel scan_data_dir param:
Unfortunatelly I can't bin it as it's used to trigger a rescan of the
Data dir when the panel is first shown. Readability got a bit better
but this is still a WIP - fullRefresh in particullar triggers a crc
recalculation but I am reluctant to rename it as the skips stuff is also
still WIP and may mess with that.