MelBase.loadData (type -> sub_type):
Mopy/bash/brec.py:
The only probable real use of "type" parameter is in MelGroups
(and MelGroup):

-        if type == self.type0:
+        if sub_type == self.type0:

-        self.loaders[type].loadData(target,ins,type,size,readId)
+        self.loaders[sub_type].loadData(target, ins, sub_type, size, readId)

MelStruct.loadData does not use its type param and in MelXpci the value
was never used - renamed to sub_type_. Similar remarks apply to the rest
of the record files.

Took the opportunity to rename local variable type -> subType in
getLoaders().