Breaking import cycle (RecordHeader to <game>/init.py): RRR
Finally noticed that RecordHeader is not used in records.py and can
be moved. This got me rid of the last of `import *` (apart the ctypes
ones and the ones in links.py which are still WIP).
`from .` imports are evil - see `from . import esp` gone (yey)
Took the opportunity to wrap a bit.

Under 65.