[!]Change - espms crcs are _no more_ recalculated in simple projects:
rsDir won't be empty for complex project's subpackages while it would
be for simple projects - so the crc would be recalculated for those.
This is inconsistent - now crcs won't be recalculated but for espms
in Data/ dir - makes sense as changes in the data dir can leave size and
mod time unchanged (for instance flip the esm bit). User should change
at least mod time if manually editing files in the project's dir.