[!]dropping _gpath:
Does not seem to cause problems the past 3 months that lays on my branch.
Asserting however `data` dicts keys are Paths would be great (except
MastersList where it is the mi and people list where it's unicode).
I have an ASSERT commit on my wip branch which however needs work.