Bye Tank - cures flickering in mods tab:
Using named tuple for fonts (will come handy)
New API for setting the UI applied in ex Tank subclasses and ModList.
Mouse texts and item formatting must be set together and the tuple return
value of getGuiKeys clearly suggested a class to group the format
elements. This finally opens the way to add more mouse texts etc in BAIN.
The flickering is cured with this commit, for reasons unknown, but time
will tell.

Under #163.
Closes #179.

Yey!