Pruning _gList usages outside of UIList:
Made __setUI private - _gList must become private to UIList finally as
a first step towards #255.

Seems that one can't DnD while clicking on the icon so this was not
needed - deleted now unused method SetDnD().

Took the opportunity to prune another use of _gList - the override was
not needed (and slightly buggy, `if selected > 0:` is always true, even
for empty list). See:

http://wxpython-users.1045709.n5.nabble.com/how-to-find-onto-all-the-items-of-a-ListCtrl-td2355029.html

for FindItem, replaced by the use of ListCtrl API
Under renames (#259) check rename links.