POC new api to fix override skips:
Fixes bug reported in #270 - basically if a package has override skips
on and data refresh skips the files the package installs they will be
reported as missing even if there (etc). I had to introduce an
overridden_skips cache populated while loading saved installers data -
just before the scan of the data dir is performed - so they can be
included in InstallersData.data_sizeCrcDate. This cache is also
populated when the override skips flag on the installer is turned on,
in case the user requests a manual Refresh Data, so the overridden skips
are not again skipped. It is _not_ however emptied (but on a restart, to
be populated again while loading Installers.dat). This is in par with
the behavior of turning skips on - the skipped files still show as dirty
till a BAIN restart that BAIN may loose track of them. The only
inconsistency is however that on manual RefreshData the files from skips
are indeed lost (if they belong to skipped Data dirs like distantlod)
but the overridden skips for a package overridden skips were turned off
still are detected. This is barely worth the performance (memory and
otherwise) to update the cache on deletion (note there may be overrides
between packages that have still override skips on).

Mopy/bash/basher/installer_links.py:
Cleanup refresh, IIUC enabling/disabling overrideSkips or hasExtraData
may change installers status for any installer, so I always refresh 'S'.
I see no reason why I should update_for_overridden_skips in the extra
data case.
Fixed a regression in the help message of skip voices introduced here:
6c2b5ef4f5dc91277633ded7cfc8a355cbdbf43b