delete_Refresh minor fixup and WIP:
doRefresh was not popped - innocuous as InstallersList is using the
_shellUI path, a bug no less.
We should probably add processing (update mods/ini tables here ?
- see #279)
Markers are already popped but now I let irefresh pop the installers.
The 'O' part is needed for delete (and added) stuff but since deleted
is passed to irefresh _all_ refreshes are triggered - see #205.
We need to split pending in added and updated (or keep pending but add
added). The `if changed` test is too broad - and I am not sure I got
all the branches. WIP