Nightlies version special case:
To allow for my 307.201603171733 betas naming scheme. This way I can
copy paste this version and the script will properly and uniquely
split it but also the version in bass can be handled by the code
which expects a float (in the wizard check for version) - closes  #285.

Not final (note that still args.version is used in filenames - not a bad
thing per se) but getting there.

Under #235, #260.