Proper fix for installer adding empty dirs in standalone:
See http://stackoverflow.com/a/36091054/281545.
That fixes the template too - I would squash with those commits

This

-                        /x "${WB_CLEAN_MOPY}\bash\basher\*.*" \
-                        /x "${WB_CLEAN_MOPY}\bash\bosh\*.*" \
-                        /x "${WB_CLEAN_MOPY}\bash\chardet\*.*" \
-                        /x "${WB_CLEAN_MOPY}\bash\game\*.*" \
-                        /x "${WB_CLEAN_MOPY}\bash\patcher\*.*" \

did not work - not sure what's the proper syntax, so still a hack - plus
as noted in the link I was not able to pin the exclusion pattern for
just the dirs I needed.