Slight refactoring:
The idea was to print packed files sorted but apparently 7z packs them
by filetype (makes sense) so the order in the list file is not kept.
I left the sorted call however - the list file is sorted.
x.path was a str - would be fun to commit a unicode path and see if this
blows, I hope not. all_files is all files we wish to pack, made it
more clear.