Tracking uses of getTweakFileSettings - WIP, TODOs:
Used in:

- IniFile.getTweakFileLines: see next commits, has also OBSEIniFile
override

The INIInfo uses are trickier cause it's called on the iniInfos.ini.
We need to associate an ini file with iniInfo instances but what we
do when iniInfos.ini changes ?

- INIInfo.getStatus:
Optimizing getStatus() out is a very laudable practice. By the time we
call set_item_format or we display a menu the ini info getStatus should
have been called except if TARGET changed - note that new infos
instances are created in refresh so their _status is None triggering
getStatus inside set_item_format. But for already existing and unchanged
infos we should not call getStatus - except if TARGET changed - or
created in the case of the game ini - needs work.
- INIInfo.listErrors: we need an ini file instance associated with
the ini info instance.

Dropped the mtimes checks for the game ini from SaveInfos as now
getSettings performs those to update INIFile cache.

TODO: We should avoid copying the cache in getSetting - factor out
_cache updating code.

Under #247.